Update resources.cpp
* size_t/int

+++ src/resources.cpp
@@ -31,9 +31,9 @@ bool ResourceProvider::searchMatch(const string name, const string searchPattern
         int offset = name.find(parts[n], pos);
         if (offset < 0)
             return false;
-        pos = offset + parts[n].length();
+        pos = offset + static_cast<int>(parts[n].length());
     }
-    return pos == int(name.length());
+    return pos == static_cast<int>(name.length());
 }
 
 string ResourceStream::readLine()
